test(heartbeat): isolate structured delivery regressions
